# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "dovecote",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "error_description": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_description"
        ]
      },
...
      "TokenIssueResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "tokenId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": "number"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "tokenId",
          "userId",
          "scopes",
          "expiresAt"
        ]
      },
...
      "TokenExchangeRequest": {
        "type": "object",
        "properties": {
          "expiresIn": {
            "type": "string",
            "enum": [
              "7d",
              "30d",
              "60d",
              "90d"
            ]
          },
          "label": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/v1/auth/exchange": {
      "post": {
        "description": "Exchange an OAuth bearer (with dovecote:admin scope) for a dvct_* runtime token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenIssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed zod validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — Retry-After header indicates retry window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin token revocation pressure circuit open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}